% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_sample.R
\name{cond_sample.default}
\alias{cond_sample.default}
\title{Method for when the user supplied distribution doesn't match one of the
options in the code.}
\usage{
\method{cond_sample}{default}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Method for when the user supplied distribution doesn't match one of the
options in the code.
}
