% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_sample.R
\name{cond_sample}
\alias{cond_sample}
\title{Generic function to call samplers}
\usage{
cond_sample(x, ...)
}
\arguments{
\item{u}{numeric vector, \eqn{0 \leq u  \leq 1} for all elements in u.}

\item{t0}{Nonnegative real number. The time the survival function is
conditioned on. \code{t0 = 0} will return samples from the unconditional distribution}

\item{HR}{hazard ratio between treatment arm and control arm}

\item{params}{named list of parameters for distribution}
}
\description{
Generic function to call samplers
}
