% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_samples.R
\name{draw_samples}
\alias{draw_samples}
\title{Draws samples from a user specified conditional survival function}
\usage{
draw_samples(n_samples, distribution, t0, HR = 1, params)
}
\arguments{
\item{n_samples}{Positive integer. The number of samples to draw from
conditional survival function}

\item{distribution}{Character vector. The name of the conditional distribution
you want to sample from (FILL in options later)}

\item{t0}{Nonnegative real number. The time the survival function is
conditioned on. \code{t0 = 0} will return samples from the unconditional distribution}

\item{HR}{hazard ratio between treatment arm and control arm}

\item{params}{named list of parameters for distribution}
}
\value{
a vector of conditional survival times from the specified distribution.
}
\description{
Draws samples from a user specified conditional survival function
}
