% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{An example function to see if documentation works. Adds two numbers together}
\usage{
example_function(x, y)
}
\arguments{
\item{x}{A real number}

\item{y}{A real number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
An example function to see if documentation works. Adds two numbers together
}
\examples{
example_function(1, 1)
example_function(3, 5)
}
