% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional_times.R
\name{get_conditional_times}
\alias{get_conditional_times}
\title{Generate samples from the posterior predictive of conditional survival times.}
\usage{
get_conditional_times(time, event, trt, distribution, n_events = NA,
  params)
}
\description{
Generate samples from the posterior predictive of conditional survival times.
}
