% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional_times.R
\name{get_single_conditional_times}
\alias{get_single_conditional_times}
\title{Generate ONE sample from the posterior predictive of conditional survival times.}
\usage{
get_single_conditional_times(time, event, trt, distribution,
  n_events = NA, params)
}
\arguments{
\item{time}{positive numeric. Survival times}

\item{event}{logical or 0/1. \code{event = 1} indicates failure,
\code{event = 0} indicates censoring}

\item{trt}{0/1 vector. Indicates assignment to treatment or control}

\item{n_events}{numeric or NULL. How many events should be observed before
censoring. If NA, there will be no censoring time.}

\item{log_HR}{log hazard ratio, usually posterior sample}
}
\description{
Get the posterior predictive distribution of conditional survival times for a
SINGLE posterior sample. This function should be interated
over all posterior samples.
}
