% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_logistic_inverse_conditional.R
\name{inverse_log_logistic}
\alias{inverse_log_logistic}
\title{Simulate random variables from a Weibull distribution. The
function is parameterized with hazard: \eqn{\nu \lambda t^{nu-1}}}
\usage{
inverse_log_logistic(u, t0, HR = 1, params)
}
\arguments{
\item{u}{numeric vector, \eqn{0 \leq u  \leq 1} for all elements in u.}

\item{t0}{Nonnegative real number. The time the survival function is
conditioned on. \code{t0 = 0} will return samples from the unconditional distribution}

\item{HR}{hazard ratio between treatment arm and control arm}

\item{params}{a named list containing \code{alpha}, the scale
parameter (and media), and \code{beta}, the shape parameter}
}
\value{
a vector of failure times conditional on \code{t0}
}
\description{
Simulate random variables from a Weibull distribution. The
function is parameterized with hazard: \eqn{\nu \lambda t^{nu-1}}
}
