% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_sample.lognormal.R
\name{log_normal_inverse_conditional}
\alias{log_normal_inverse_conditional}
\title{Draws samples from a Lognormal survival model conditional on a timepoint and adjusted for a hazard ratio, \eqn{T \sim f(t|t_0; \gamma)}}
\usage{
log_normal_inverse_conditional(u, t0, HR, params)
}
\arguments{
\item{u}{numeric vector, \eqn{0 \leq u  \leq 1} for all elements in u.}

\item{t0}{Nonnegative real number. The time the survival function is
conditioned on. \code{t0 = 0} will return samples from the unconditional distribution}

\item{HR}{hazard ratio between treatment arm and control arm}

\item{params}{a named list containing \code{mu}, mean of lognormal distribution
on the log-scale, and \code{sigma}, standard deviation of lognormal distribution on the log-scale}
}
\value{
vector of survival times following the Lognormal survival model parameterized by \eqn{mu} and \eqn{sigma}
}
\description{
Draws samples from a Lognormal survival model conditional on a timepoint and adjusted for a hazard ratio, \eqn{T \sim f(t|t_0; \gamma)}
}
\examples{
log_normal_inverse_conditional(0.5, t0 = 1, mu = 0.2, sigma = 1, gamma = 1)
log_normal_inverse_conditional(seq(0, 1, by = 0.2))
}
