% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_jags}
\alias{run_jags}
\title{Runs a user-supplied JAGS model and returns the output}
\usage{
run_jags(
  model_file,
  data,
  chains,
  iter = 10000L,
  burn = 1000L,
  track_variable_names,
  progress.bar = "none"
)
}
\arguments{
\item{model_file}{string containing name of JAGS model to run. Must be stored in jags/ for now}

\item{data}{list of parameters and data to pass to JAGS}

\item{chains}{integer number of MCMC chains to run}

\item{iter}{integer number of MCMC iterations, default 100,000}

\item{burn}{integer length of burnin, default 1,000}

\item{track_variable_names}{a character vector giving the names of variables to be monitored}

\item{progress.bar}{should JAGS show a progress bar, "none" (default_), "text" , or "GUI"}
}
\value{
a mcmc.list of posterior draws of parameters in \code{track_variable_names}, contains matrix of size (iter x track_num_variables)
}
\description{
Runs a user-supplied JAGS model and returns the output
}
