% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_sample.weibull.R
\name{cond_sample.weibull}
\alias{cond_sample.weibull}
\title{Simulate random variables from a Weibull distribution. The
function is parameterized with hazard: \eqn{\nu \lambda t^{nu-1}}}
\usage{
\method{cond_sample}{weibull}(u, t0, trt = 0, params)
}
\arguments{
\item{u}{numeric vector, \eqn{0 \leq u  \leq 1} for all elements in u.}

\item{t0}{Nonnegative real number. The time the survival function is
conditioned on. \code{t0 = 0} will return samples from the unconditional distribution}

\item{params}{a named list containing \code{lambda}, the Weibull scale
parameter, and \code{nu}, the Weibull shape parameter, and \code{log_HR},
log hazard ratio}
}
\value{
a vector of failure times conditional on \code{t0}
}
\description{
Simulate random variables from a Weibull distribution. The
function is parameterized with hazard: \eqn{\nu \lambda t^{nu-1}}
}
