% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal_inverse_conditional.R
\name{log_normal_inverse_conditional}
\alias{log_normal_inverse_conditional}
\title{Draws samples from a Lognormal survival model conditional on a timepoint and adjusted for hazard ratio}
\usage{
log_normal_inverse_conditional(u, t0 = 0, mu = 0, sigma = 1,
  gamma = 1)
}
\arguments{
\item{u}{vector of numbers between 0 and 1. Generally a vector of uniform \link{0,1} random variable}

\item{t0}{time which the survival function is conditioned on, \eqn{S(t|t0)}. When \code{t0 = 0} there is no conditioning. Default to 0}

\item{mu}{mean of lognormal distribution on the log-scale, default value of 0}

\item{sigma}{standard deviation of lognormal distribution on the log-scale, default value of 1}

\item{gamma}{hazard ratio, default value of 1}
}
\value{
vector of survival times following the Lognormal survival model parameterized by \eqn{mu} and \eqn{sigma}
}
\description{
Draws samples from a Lognormal survival model conditional on a timepoint and adjusted for hazard ratio
}
\examples{
log_normal_inverse_conditional(0.5, t0 = 1, mu = 0.2, sigma = 1, gamma = 1)
log_normal_inverse_conditional(seq(0, 1, by = 0.2))
}
