% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_jags}
\alias{run_jags}
\title{Runs a user-supplied JAGS model and returns the output}
\usage{
run_jags(distribution, prop_haz = TRUE, time, event, trt, chains,
  iter = 10000L, burn = 1000L, n_adapt = 1000L, track_variable_names,
  progress.bar = "none")
}
\arguments{
\item{distribution}{Character. name of the distribution of the parametric model}

\item{prop_haz}{logical. Should a proportional hazards model be fit (TRUE) or a
nonproportional hazards model}

\item{time}{Positive numeric. survival times}

\item{event}{0/1 variable. Event happened (\code{event = 1}) or observation is
censored (\code{event = 0})}

\item{trt}{0/1 variable. Indicates control arm (\code{trt = 0}) or
experimental arm (\code{trt = 1})}

\item{chains}{integer number of MCMC chains to run}

\item{iter}{integer number of MCMC iterations, default 100,000}

\item{burn}{integer length of burn in, default 1,000}

\item{track_variable_names}{a character vector giving the names of variables to be monitored}

\item{progress.bar}{should JAGS show a progress bar, "none" (default_), "text" , or "GUI"}
}
\value{
a mcmc.list of posterior draws of parameters in \code{track_variable_names}, contains matrix of size (iter x track_num_variables)
}
\description{
Runs a user-supplied JAGS model and returns the output
}
